refactor: extract manager to different file
